snapshot machinery
