+++ storage/deposit/merkle/README.md
@@ -1,0 +1,5 @@
+# Deposit Merkle Tree
+
+This package implements the [EIP-4881 spec](https://eips.ethereum.org/EIPS/eip-4881) for a minimal sparse Merkle tree.
+
+The format proposed by the EIP allows for the pruning of deposits that are no longer needed to participate fully in consensus.
