+++ storage/deposit/merkle/constants.go
@@ -1,4 +1,4 @@
 package merkle
 
 // DepositContractDepth is the maximum tree depth as defined by EIP-4881.
-const DepositContractDepth = 32
+const DepositContractDepth uint64 = 32
