+++ storage/deposit/merkle/tree.go
@@ -1,309 +1,158 @@
 package merkle
 
 import (
-	"slices"
+	"encoding/binary"
 
-	"github.com/berachain/beacon-kit/primitives/math/pow"
+	"github.com/berachain/beacon-kit/errors"
+	"github.com/berachain/beacon-kit/primitives/crypto/sha256"
+	"github.com/berachain/beacon-kit/primitives/math"
 	"github.com/berachain/beacon-kit/primitives/merkle"
-	"github.com/berachain/beacon-kit/primitives/merkle/zero"
+	"github.com/ethereum/go-ethereum/common"
 )
 
-// create builds a new merkle tree
-func create(hasher merkle.Hasher[[32]byte], leaves [][32]byte, depth uint64) TreeNode {
-	length := uint64(len(leaves))
-	if length == 0 {
-		return &ZeroNode{
-			depth:  depth,
-			hasher: hasher,
-		}
-	}
+// DepositTree is the Merkle tree representation of deposits.
+type DepositTree struct {
+	tree                    TreeNode
+	mixInLength             uint64
+	finalizedExecutionBlock executionBlock
+	hasher                  merkle.Hasher[[32]byte]
+}
 
-	if depth == 0 {
-		return &LeafNode{hash: leaves[0]}
-	}
+type executionBlock struct {
+	Hash  [32]byte
+	Depth uint64
+}
 
-	split := min(pow.TwoToThePowerOf(depth-1), length)
-	left := create(hasher, leaves[0:split], depth-1)
-	right := create(hasher, leaves[split:], depth-1)
-	return &InnerNode{
-		left:   left,
-		right:  right,
-		hasher: hasher,
+// NewDepositTree creates an empty deposit tree.
+func NewDepositTree() *DepositTree {
+	var (
+		hasher = merkle.NewHasher[[32]byte](sha256.Hash)
+		leaves [][32]byte
+	)
+	merkle := create(hasher, leaves, DepositContractDepth)
+	return &DepositTree{
+		tree:                    merkle,
+		mixInLength:             0,
+		finalizedExecutionBlock: executionBlock{},
+		hasher:                  hasher,
 	}
 }
 
-// fromSnapshotParts creates a new Merkle tree from a list of finalized leaves, number of deposits and specified depth.
-func fromSnapshotParts(
-	hasher merkle.Hasher[[32]byte], finalized [][32]byte, deposits uint64, level uint64,
-) (TreeNode, error) {
-	var err error
+// GetSnapshot returns a deposit tree snapshot.
+func (d *DepositTree) GetSnapshot() (DepositTreeSnapshot, error) {
+	var finalized [][32]byte
+	mixInLength, finalized := d.tree.GetFinalized(finalized)
+	return fromTreeParts(finalized, mixInLength, d.finalizedExecutionBlock)
+}
 
-	if len(finalized) < 1 || deposits == 0 {
-		return &ZeroNode{
-			depth:  level,
-			hasher: hasher,
-		}, nil
+// Finalize marks a deposit as finalized.
+func (d *DepositTree) Finalize(eth1DepositIndex int64, executionHash common.Hash, executionNumber uint64) error {
+	var blockHash [32]byte
+	copy(blockHash[:], executionHash[:])
+	d.finalizedExecutionBlock = executionBlock{
+		Hash:  blockHash,
+		Depth: executionNumber,
 	}
-	if deposits == pow.TwoToThePowerOf(level) {
-		return &FinalizedNode{
-			depositCount: deposits,
-			hash:         finalized[0],
-		}, nil
+	mixInLength := uint64(eth1DepositIndex + 1)
+	_, err := d.tree.Finalize(mixInLength, DepositContractDepth)
+	if err != nil {
+		return err
 	}
-	if level == 0 {
-		return nil, ErrZeroLevel
+	return nil
+}
+
+// getProof returns the deposit tree proof.
+func (d *DepositTree) getProof(index uint64) ([32]byte, [][32]byte, error) {
+	if d.mixInLength <= 0 {
+		return [32]byte{}, nil, ErrInvalidDepositCount
 	}
-	node := InnerNode{
-		hasher: hasher,
+	if index >= d.mixInLength {
+		return [32]byte{}, nil, ErrInvalidIndex
 	}
-	if leftSubtree := pow.TwoToThePowerOf(level - 1); deposits <= leftSubtree {
-		node.left, err = fromSnapshotParts(hasher, finalized, deposits, level-1)
+	finalizedDeposits, _ := d.tree.GetFinalized([][32]byte{})
+	finalizedIdx := -1
+	if finalizedDeposits != 0 {
+		fd, err := math.Int(finalizedDeposits)
 		if err != nil {
-			return nil, err
+			return [32]byte{}, nil, err
 		}
-		node.right = &ZeroNode{
-			depth:  level - 1,
-			hasher: hasher,
-		}
-	} else {
-		node.left = &FinalizedNode{
-			depositCount: leftSubtree,
-			hash:         finalized[0],
-		}
-		node.right, err = fromSnapshotParts(hasher, finalized[1:], deposits-leftSubtree, level-1)
-		if err != nil {
-			return nil, err
-		}
+		finalizedIdx = fd - 1
 	}
-	return &node, nil
-}
-
-// generateProof returns a merkle proof and root
-func generateProof(tree TreeNode, index uint64, depth uint64) ([32]byte, [][32]byte) {
-	var proof [][32]byte
-	node := tree
-	for depth > 0 {
-		ithBit := (index >> (depth - 1)) & 0x1
-		if ithBit == 1 {
-			proof = append(proof, node.Left().GetRoot())
-			node = node.Right()
-		} else {
-			proof = append(proof, node.Right().GetRoot())
-			node = node.Left()
-		}
-		depth--
+	i, err := math.Int(index)
+	if err != nil {
+		return [32]byte{}, nil, err
 	}
+	if finalizedDeposits > 0 && i <= finalizedIdx {
+		return [32]byte{}, nil, ErrInvalidIndex
+	}
+	leaf, proof := generateProof(d.tree, index, DepositContractDepth)
 
-	slices.Reverse(proof)
-	return node.GetRoot(), proof
+	mixInLength := [32]byte{}
+	binary.LittleEndian.PutUint64(mixInLength[:], d.mixInLength)
+	proof = append(proof, mixInLength)
+	return leaf, proof, nil
 }
 
-// FinalizedNode represents a finalized node and satisfies the TreeNode interface.
-type FinalizedNode struct {
-	depositCount uint64
-	hash         [32]byte
+// getRoot returns the root of the deposit tree.
+func (d *DepositTree) getRoot() [32]byte {
+	var enc [32]byte
+	binary.LittleEndian.PutUint64(enc[:], d.mixInLength)
+
+	root := d.tree.GetRoot()
+	return d.hasher.Combi(root, enc)
 }
 
-// GetRoot returns the root of the Merkle tree.
-func (f *FinalizedNode) GetRoot() [32]byte {
-	return f.hash
-}
-
-// IsFull returns whether there is space left for deposits.
-// A FinalizedNode will always return true as by definition it
-// is full and deposits can't be added to it.
-func (*FinalizedNode) IsFull() bool {
-	return true
-}
-
-// Finalize marks deposits of the Merkle tree as finalized.
-func (f *FinalizedNode) Finalize(depositsToFinalize uint64, depth uint64) (TreeNode, error) {
-	return f, nil
-}
-
-// GetFinalized returns a list of hashes of all the finalized nodes and the number of deposits.
-func (f *FinalizedNode) GetFinalized(result [][32]byte) (uint64, [][32]byte) {
-	return f.depositCount, append(result, f.hash)
-}
-
-// PushLeaf adds a new leaf node at the next available zero node.
-func (*FinalizedNode) PushLeaf([32]byte, uint64) (TreeNode, error) {
-	return nil, ErrFinalizedNodeCannotPushLeaf
-}
-
-// Right returns nil as a finalized node can't have any children.
-func (*FinalizedNode) Right() TreeNode {
-	return nil
-}
-
-// Left returns nil as a finalized node can't have any children.
-func (*FinalizedNode) Left() TreeNode {
-	return nil
-}
-
-// LeafNode represents a leaf node holding a deposit and satisfies the TreeNode interface.
-type LeafNode struct {
-	hash [32]byte
-}
-
-// GetRoot returns the root of the Merkle tree.
-func (l *LeafNode) GetRoot() [32]byte {
-	return l.hash
-}
-
-// IsFull returns whether there is space left for deposits.
-// A LeafNode will always return true as it is the last node
-// in the tree and therefore can't have any deposits added to it.
-func (*LeafNode) IsFull() bool {
-	return true
-}
-
-// Finalize marks deposits of the Merkle tree as finalized.
-func (l *LeafNode) Finalize(depositsToFinalize uint64, depth uint64) (TreeNode, error) {
-	return &FinalizedNode{1, l.hash}, nil
-}
-
-// GetFinalized returns a list of hashes of all the finalized nodes and the number of deposits.
-func (*LeafNode) GetFinalized(result [][32]byte) (uint64, [][32]byte) {
-	return 0, result
-}
-
-// PushLeaf adds a new leaf node at the next available zero node.
-func (*LeafNode) PushLeaf([32]byte, uint64) (TreeNode, error) {
-	return nil, ErrLeafNodeCannotPushLeaf
-}
-
-// Right returns nil as a leaf node is the last node and can't have any children.
-func (*LeafNode) Right() TreeNode {
-	return nil
-}
-
-// Left returns nil as a leaf node is the last node and can't have any children.
-func (*LeafNode) Left() TreeNode {
-	return nil
-}
-
-// InnerNode represents an inner node with two children and satisfies the TreeNode interface.
-type InnerNode struct {
-	left, right TreeNode
-	hasher      merkle.Hasher[[32]byte]
-}
-
-// GetRoot returns the root of the Merkle tree.
-func (n *InnerNode) GetRoot() [32]byte {
-	left := n.left.GetRoot()
-	right := n.right.GetRoot()
-	return n.hasher.Combi(left, right)
-}
-
-// IsFull returns whether there is space left for deposits.
-func (n *InnerNode) IsFull() bool {
-	return n.right.IsFull()
-}
-
-// Finalize marks deposits of the Merkle tree as finalized.
-func (n *InnerNode) Finalize(depositsToFinalize uint64, depth uint64) (TreeNode, error) {
+// pushLeaf adds a new leaf to the tree.
+func (d *DepositTree) pushLeaf(leaf [32]byte) error {
 	var err error
-	deposits := pow.TwoToThePowerOf(depth)
-	if deposits <= depositsToFinalize {
-		return &FinalizedNode{deposits, n.GetRoot()}, nil
+	d.tree, err = d.tree.PushLeaf(leaf, DepositContractDepth)
+	if err != nil {
+		return err
 	}
-	if depth == 0 {
-		return nil, ErrZeroDepth
+	d.mixInLength++
+	return nil
+}
+
+// Insert is defined as part of MerkleTree interface and adds a new leaf to the tree.
+func (d *DepositTree) Insert(item []byte, _ int) error {
+	var leaf [32]byte
+	copy(leaf[:], item[:32])
+	return d.pushLeaf(leaf)
+}
+
+// HashTreeRoot is defined as part of MerkleTree interface and calculates the hash tree root.
+func (d *DepositTree) HashTreeRoot() ([32]byte, error) {
+	root := d.getRoot()
+	if root == [32]byte{} {
+		return [32]byte{}, errors.New("could not retrieve hash tree root")
 	}
-	n.left, err = n.left.Finalize(depositsToFinalize, depth-1)
+	return root, nil
+}
+
+// NumOfItems is defined as part of MerkleTree interface and returns the number of deposits in the tree.
+func (d *DepositTree) NumOfItems() int {
+	return int(d.mixInLength)
+}
+
+// MerkleProof is defined as part of MerkleTree interface and generates a merkle proof.
+func (d *DepositTree) MerkleProof(index int) ([][]byte, error) {
+	_, proof, err := d.getProof(uint64(index))
 	if err != nil {
 		return nil, err
 	}
-	if depositsToFinalize > deposits/2 {
-		remaining := depositsToFinalize - deposits/2
-		n.right, err = n.right.Finalize(remaining, depth-1)
-		if err != nil {
-			return nil, err
-		}
+	byteSlices := make([][]byte, len(proof))
+	for i, p := range proof {
+		copied := p
+		byteSlices[i] = copied[:]
 	}
-	return n, nil
+	return byteSlices, nil
 }
 
-// GetFinalized returns a list of hashes of all the finalized nodes and the number of deposits.
-func (n *InnerNode) GetFinalized(result [][32]byte) (uint64, [][32]byte) {
-	leftDeposits, result := n.left.GetFinalized(result)
-	rightDeposits, result := n.right.GetFinalized(result)
-	return leftDeposits + rightDeposits, result
-}
-
-// PushLeaf adds a new leaf node at the next available zero node.
-func (n *InnerNode) PushLeaf(leaf [32]byte, depth uint64) (TreeNode, error) {
-	if !n.left.IsFull() {
-		left, err := n.left.PushLeaf(leaf, depth-1)
-		if err == nil {
-			n.left = left
-		} else {
-			return n, err
-		}
-	} else {
-		right, err := n.right.PushLeaf(leaf, depth-1)
-		if err == nil {
-			n.right = right
-		} else {
-			return n, err
-		}
+// Copy performs a deep copy of the tree.
+func (d *DepositTree) Copy() (*DepositTree, error) {
+	snapshot, err := d.GetSnapshot()
+	if err != nil {
+		return nil, err
 	}
-	return n, nil
-}
-
-// Right returns the child node on the right.
-func (n *InnerNode) Right() TreeNode {
-	return n.right
-}
-
-// Left returns the child node on the left.
-func (n *InnerNode) Left() TreeNode {
-	return n.left
-}
-
-// ZeroNode represents an empty node without a deposit and satisfies the TreeNode interface.
-type ZeroNode struct {
-	depth  uint64
-	hasher merkle.Hasher[[32]byte]
-}
-
-// GetRoot returns the root of the Merkle tree.
-func (z *ZeroNode) GetRoot() [32]byte {
-	if z.depth == DepositContractDepth {
-		return z.hasher.Combi(zero.Hashes[z.depth-1], zero.Hashes[z.depth-1])
-	}
-	return zero.Hashes[z.depth]
-}
-
-// IsFull returns wh   ether there is space left for deposits.
-// A ZeroNode will always return false as a ZeroNode is an empty node
-// that gets replaced by a deposit.
-func (*ZeroNode) IsFull() bool {
-	return false
-}
-
-// Finalize marks deposits of the Merkle tree as finalized.
-func (*ZeroNode) Finalize(depositsToFinalize uint64, depth uint64) (TreeNode, error) {
-	return nil, nil
-}
-
-// GetFinalized returns a list of hashes of all the finalized nodes and the number of deposits.
-func (*ZeroNode) GetFinalized(result [][32]byte) (uint64, [][32]byte) {
-	return 0, result
-}
-
-// PushLeaf adds a new leaf node at the next available zero node.
-func (z *ZeroNode) PushLeaf(leaf [32]byte, depth uint64) (TreeNode, error) {
-	return create(z.hasher, [][32]byte{leaf}, depth), nil
-}
-
-// Right returns nil as a zero node can't have any children.
-func (*ZeroNode) Right() TreeNode {
-	return nil
-}
-
-// Left returns nil as a zero node can't have any children.
-func (*ZeroNode) Left() TreeNode {
-	return nil
+	return fromSnapshot(d.hasher, snapshot)
 }
